LibC: Make sure that ioctl() requests are #defined as macros
This fixes terminal UI resizing in the vim port. The problem was that
vim had "#ifdef TIOCGWINSZ" around the code that figures out the size
of the terminal.

Since all of our ioctl() requests were enum values, this code was not
compiled into vim at all. This patch fixes that. :^)